refactor: 역할별 패키지 재구성
