+++ src/main/java/lotto/domain/WinningLottoTicket.java
@@ -4,6 +4,8 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import lotto.domain.enums.Prize;
+
 /**
  * @author : byungkyu
  * @date : 2020/12/14
