+++ src/main/java/lotto/domain/lottoTicketsStrategy/AutoLottoTicketsStrategy.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottoTicketsStrategy;
 
 import static lotto.domain.LottoTickets.*;
 
@@ -8,7 +8,10 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import lotto.view.BuyInfo;
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoTicket;
+import lotto.domain.LottoTickets;
+import lotto.domain.BuyInfo;
 
 /**
  * @author : byungkyu
