+++ src/main/java/lotto/domain/lottoTicketsStrategy/ManualLottoTicketsStrategy.java
@@ -1,10 +1,14 @@
-package lotto.domain;
+package lotto.domain.lottoTicketsStrategy;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import lotto.view.BuyInfo;
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoTicket;
+import lotto.domain.LottoTickets;
+import lotto.domain.ManualNumbers;
+import lotto.domain.BuyInfo;
 
 /**
  * @author : byungkyu
