+++ src/test/java/lotto/domain/WinningStatisticsTest.java
@@ -9,6 +9,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import lotto.domain.enums.Prize;
+
 /**
  * @author : byungkyu
  * @date : 2020/12/14
